requirements.txt
Flask
joblib
jsonschema
marshmallow
requests
bugsnag[flask]
pendulum
cerberus
marshmallow
mysqlclient
flask-jwt-extended
flask-migrate
flask-sqlalchemy
numpy